django_admin/requirements/base.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-model-utils==4.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1

